Extract staged loading into ord_schema.orm.loading; dispose worker engines
Move the staged/parallel loading orchestration out of the add_datasets script and
into a library module so consumers (e.g. ord-interface) import it from the library
rather than reaching into scripts/.

- New ord_schema/orm/loading.py holds load_datasets() plus the per-dataset helpers
  (ingest_dataset, derive_dataset, add_rdkit, dataset_id_for_file). add_datasets.py
  is now a thin CLI: argument parsing, stage/extra validation, and a delegating
  main().
- Each worker now disposes the engine it creates (try/finally) so pooled connections
  don't accumulate in reused ProcessPoolExecutor workers; the serial RDKit engine is
  disposed too.
- _run_parallel returns (results, failures) instead of mutating a caller-owned list.
- Tests: loading behavior moves to ord_schema/orm/loading_test.py (calls
  load_datasets directly); add_datasets_test.py keeps the CLI-level cases; the shared
  prepared_engine fixture moves to the orm conftest.
