+++ ord_schema/orm/loading.py
@@ -1,0 +1,277 @@
+# Copyright 2022 Open Reaction Database Project Authors
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""Staged loading of ORD datasets into the ORM database.
+
+Loading is two independent stages: *ingest* writes the ``ord.*`` search index and
+``public.*`` payload, and *derivation* writes the ``derived.*`` SMILES, RDKit links,
+and (optionally) reaction classes. Either stage can run without the other; derivation
+is idempotent, so the derived-only stage backfills or recomputes derived data over
+already-ingested datasets. ``load_datasets`` orchestrates both stages over a glob of
+dataset files; the per-dataset helpers (``ingest_dataset``, ``derive_dataset``,
+``add_rdkit``) are exposed for callers that compose their own pipeline.
+"""
+
+import time
+from collections.abc import Callable, Iterable
+from concurrent.futures import ProcessPoolExecutor, ThreadPoolExecutor, as_completed
+from contextlib import ExitStack
+from functools import partial
+from glob import glob
+from hashlib import md5
+
+from sqlalchemy import create_engine
+from sqlalchemy.engine import Engine
+from sqlalchemy.orm import Session
+from tqdm import tqdm
+
+from ord_schema import parquet
+from ord_schema.logging import get_logger
+from ord_schema.message_helpers import load_message
+from ord_schema.orm import database
+from ord_schema.proto import dataset_pb2
+
+logger = get_logger(__name__)
+
+STAGES = ("ingest", "derived")
+
+
+def dataset_id_for_file(filename: str) -> str:
+    """Returns the dataset ID for a dataset file without ingesting it."""
+    if filename.endswith(".parquet"):
+        return parquet.load_metadata(filename).dataset_id
+    return load_message(filename, dataset_pb2.Dataset).dataset_id
+
+
+def ingest_dataset(filename: str, *, dsn: str, overwrite: bool) -> str:
+    """Ingests a single dataset (``ord.*`` + ``public.*``); skips unchanged datasets.
+
+    Derived data is populated separately by ``derive_dataset``/``add_rdkit``.
+
+    Args:
+        filename: Dataset filename. ``.parquet`` inputs are streamed via
+            ``database.add_parquet_dataset``; other formats load the whole
+            Dataset proto and call ``database.add_dataset``.
+        dsn: Database connection string.
+        overwrite: If True, update the dataset if the MD5 hash has changed.
+
+    Returns:
+        Dataset ID.
+
+    Raises:
+        ValueError: If the dataset already exists in the database and `overwrite` is not set.
+    """
+    if filename.endswith(".parquet"):
+        logger.debug(f"Streaming {filename}")
+        dataset_id = parquet.load_metadata(filename).dataset_id
+
+        def compute_md5() -> str:
+            md5_hex, _ = parquet.streaming_md5(filename)
+            return md5_hex
+
+        def insert(session: Session) -> None:
+            database.add_parquet_dataset(filename, session)
+    else:
+        logger.debug(f"Loading {filename}")
+        dataset = load_message(filename, dataset_pb2.Dataset)
+        dataset_id = dataset.dataset_id
+
+        def compute_md5() -> str:
+            return md5(
+                dataset.SerializeToString(deterministic=True), usedforsecurity=False
+            ).hexdigest()
+
+        def insert(session: Session) -> None:
+            database.add_dataset(dataset, session)
+
+    # NOTE(skearnes): Multiprocessing is hard to get right for shared connection pools, so we don't even try; see
+    # https://docs.sqlalchemy.org/en/20/core/pooling.html#using-connection-pools-with-multiprocessing-or-os-fork.
+    # Each call owns its engine and disposes it so pooled connections don't accumulate in reused pool workers.
+    engine = create_engine(dsn)
+    try:
+        with Session(engine) as session:
+            with session.begin():
+                existing_md5 = database.get_dataset_md5(dataset_id, session)
+            if existing_md5 is not None:
+                if compute_md5() != existing_md5:
+                    if not overwrite:
+                        raise ValueError(
+                            f"`overwrite` is required when a dataset already exists: {dataset_id}"
+                        )
+                    logger.debug(f"existing dataset {dataset_id} changed; updating")
+                    with session.begin():
+                        database.delete_dataset(dataset_id, session)
+                else:
+                    logger.debug(f"existing dataset {dataset_id} unchanged; skipping")
+                    return dataset_id
+            start = time.time()
+            with session.begin():
+                insert(session)
+            logger.debug(f"insert took {time.time() - start:g}s")
+    finally:
+        engine.dispose()
+    return dataset_id
+
+
+def derive_dataset(dataset_id: str, *, dsn: str, classify_reactions: bool) -> str:
+    """Runs the parallel-safe derived passes (SMILES + optional classification).
+
+    The RDKit linking pass is excluded here and run serially in ``add_rdkit`` to avoid
+    deadlocks. Idempotent, so it is safe to re-run over already-derived datasets.
+
+    Args:
+        dataset_id: Dataset to derive.
+        dsn: Database connection string.
+        classify_reactions: Whether to assign reaction class/name labels.
+
+    Returns:
+        Dataset ID.
+    """
+    engine = create_engine(dsn)
+    try:
+        with Session(engine) as session, session.begin():
+            database.update_derived_data(
+                dataset_id,
+                session,
+                rdkit_cartridge=False,  # Done serially in add_rdkit() to avoid deadlocks.
+                classify_reactions=classify_reactions,
+            )
+    finally:
+        engine.dispose()
+    return dataset_id
+
+
+def add_rdkit(engine: Engine, dataset_id: str) -> None:
+    """Populates and links the RDKit cartridge tables for a dataset."""
+    with Session(engine) as session:
+        with session.begin():
+            database.update_rdkit_tables(dataset_id, session)
+        with session.begin():
+            database.update_rdkit_ids(dataset_id, session)
+
+
+def _run_parallel(
+    func: Callable[[str], str],
+    items: Iterable[str],
+    *,
+    n_jobs: int,
+    desc: str,
+) -> tuple[list[str], list[str]]:
+    """Runs ``func`` over ``items`` with ``n_jobs`` workers.
+
+    Returns ``(results, failures)``: the successful return values and the inputs whose
+    call raised.
+    """
+    items = list(items)
+    results: list[str] = []
+    failures: list[str] = []
+    with ExitStack() as stack:
+        if n_jobs > 1:
+            executor = stack.enter_context(ProcessPoolExecutor(n_jobs))
+        else:
+            executor = stack.enter_context(ThreadPoolExecutor(n_jobs))
+        futures = {executor.submit(func, item): item for item in items}
+        for future in tqdm(as_completed(futures), total=len(futures), desc=desc):
+            try:
+                results.append(future.result())
+            except Exception:
+                item = futures[future]
+                failures.append(item)
+                logger.exception(f"{desc} failed for {item}")
+    return results, failures
+
+
+def load_datasets(
+    pattern: str,
+    dsn: str,
+    *,
+    stages: Iterable[str] = STAGES,
+    overwrite: bool = False,
+    classify_reactions: bool = False,
+    n_jobs: int = 1,
+) -> None:
+    """Runs the selected stages over the datasets matching ``pattern``.
+
+    The parallel-safe work (ingest, SMILES/classification) runs in a worker pool; the
+    RDKit linking pass runs serially to avoid deadlocks. When ingest is skipped, dataset
+    IDs are resolved from ``pattern`` so derivation runs over the already-ingested rows.
+
+    Args:
+        pattern: Glob for dataset filenames.
+        dsn: Database connection string.
+        stages: Stages to run (any of ``STAGES``); defaults to all.
+        overwrite: If True, update changed datasets during ingest.
+        classify_reactions: If True, assign reaction class/name labels during derivation.
+        n_jobs: Number of parallel workers.
+
+    Raises:
+        ValueError: If ``stages`` is empty or names an unknown stage.
+        ImportError: If ``classify_reactions`` is set without the ``reaction-class`` extra.
+        RuntimeError: If any dataset failed in any stage; the inputs are the message.
+    """
+    stages = tuple(stages)
+    unknown = set(stages) - set(STAGES)
+    if unknown:
+        raise ValueError(
+            f"unknown stages {sorted(unknown)}; choose from {list(STAGES)}"
+        )
+    if not stages:
+        raise ValueError(f"stages must include at least one of {list(STAGES)}")
+    if classify_reactions and database.update_reaction_classes is None:
+        raise ImportError(
+            "reaction classification requires the 'reaction-class' extra: "
+            "pip install ord-schema[reaction-class]"
+        )
+    filenames = sorted(glob(pattern))
+    failures: list[str] = []
+
+    if "ingest" in stages:
+        logger.info("Ingesting datasets")
+        dataset_ids, stage_failures = _run_parallel(
+            partial(ingest_dataset, dsn=dsn, overwrite=overwrite),
+            filenames,
+            n_jobs=n_jobs,
+            desc="Ingest",
+        )
+        failures.extend(stage_failures)
+    else:
+        dataset_ids = [dataset_id_for_file(filename) for filename in filenames]
+
+    if "derived" in stages:
+        logger.info("Deriving SMILES and reaction classes")
+        dataset_ids, stage_failures = _run_parallel(
+            partial(derive_dataset, dsn=dsn, classify_reactions=classify_reactions),
+            dataset_ids,
+            n_jobs=n_jobs,
+            desc="Derive",
+        )
+        failures.extend(stage_failures)
+        logger.info("Adding RDKit functionality")
+        engine = create_engine(dsn)
+        try:
+            for dataset_id in tqdm(dataset_ids, desc="RDKit"):
+                try:
+                    add_rdkit(
+                        engine, dataset_id
+                    )  # NOTE(skearnes): Do this serially to avoid deadlocks.
+                except Exception:
+                    failures.append(dataset_id)
+                    logger.exception(
+                        f"Adding RDKit functionality for {dataset_id} failed"
+                    )
+        finally:
+            engine.dispose()
+
+    if failures:
+        raise RuntimeError(failures)
