+++ ord_schema/orm/scripts/add_datasets_test.py
@@ -12,139 +12,46 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Tests for ord_schema.orm.scripts.add_datasets."""
+"""Tests for ord_schema.orm.scripts.add_datasets.
+
+Loading behavior is covered in ord_schema.orm.loading_test; these cover the CLI wiring
+(argument parsing, stage validation) and that main delegates to loading.load_datasets.
+"""
 
 import pathlib
 
 import pytest
-from sqlalchemy import select, text
-from sqlalchemy.engine import Engine
 from sqlalchemy.orm import Session
 
-from ord_schema import message_helpers, parquet
-from ord_schema.orm import database
-from ord_schema.orm.database import prepare_database
 from ord_schema.orm.mappers import Mappers
 from ord_schema.orm.scripts import add_datasets
-from ord_schema.proto import dataset_pb2, reaction_pb2
 
 _PBTXT_FIXTURE = str(
     pathlib.Path(__file__).parent / ".." / "testdata" / "ord-nielsen-example.pbtxt"
 )
 
 
-@pytest.fixture(name="prepared_engine")
-def prepared_engine_fixture(test_engine) -> Engine:
-    """``test_engine`` with the ORM schema (and RDKit cartridge) installed."""
-    assert prepare_database(test_engine)
-    return test_engine
-
-
-def _write_parquet_dataset(tmp_path) -> tuple[str, dataset_pb2.Dataset]:
-    dataset = message_helpers.load_message(_PBTXT_FIXTURE, dataset_pb2.Dataset)
-    parquet_path = (tmp_path / "dataset.parquet").as_posix()
-    parquet.save_dataset(dataset, parquet_path)
-    return parquet_path, dataset
-
-
 def test_main(prepared_engine):
+    """main() parses args and runs both stages over the matching datasets."""
     argv = ["--dsn", str(prepared_engine.url), "--pattern", _PBTXT_FIXTURE]
     add_datasets.main(add_datasets.parse_args(argv))
-
-
-def test_main_parquet(prepared_engine, tmp_path):
-    """Streaming Parquet ingest stores the streaming MD5, reaction count, and Reaction rows."""
-    parquet_path, dataset = _write_parquet_dataset(tmp_path)
-    argv = ["--dsn", str(prepared_engine.url), "--pattern", parquet_path]
-    add_datasets.main(add_datasets.parse_args(argv))
-    expected_md5, expected_count = parquet.streaming_md5(parquet_path)
-    with Session(prepared_engine) as session:
-        # Dataset row's metadata columns reflect the streaming MD5/count.
-        assert database.get_dataset_md5(dataset.dataset_id, session) == expected_md5
-        assert database.get_dataset_size(dataset.dataset_id, session) == expected_count
-        # Reaction rows were actually inserted (not just the num_reactions column).
-        mapped_dataset = session.scalar(
-            select(Mappers.Dataset).where(
-                Mappers.Dataset.dataset_id == dataset.dataset_id
-            )
-        )
-        assert mapped_dataset is not None
-        assert len(mapped_dataset.reactions) == expected_count
-        # Spot-check that one reaction round-trips byte-identical through public.reactions.
-        original = dataset.reactions[0]
-        stored = next(
-            r for r in mapped_dataset.reactions if r.reaction_id == original.reaction_id
-        )
-        assert reaction_pb2.Reaction.FromString(stored.proto_row.proto) == original
-
-
-def test_main_parquet_skip_unchanged(prepared_engine, tmp_path):
-    """A second invocation without --overwrite is a no-op when the MD5 matches."""
-    parquet_path, dataset = _write_parquet_dataset(tmp_path)
-    argv = ["--dsn", str(prepared_engine.url), "--pattern", parquet_path]
-    add_datasets.main(add_datasets.parse_args(argv))
-    add_datasets.main(add_datasets.parse_args(argv))
-    # Skip path must not re-insert: reaction count is unchanged after the second call.
-    _, expected_count = parquet.streaming_md5(parquet_path)
-    with Session(prepared_engine) as session:
-        mapped_dataset = session.scalar(
-            select(Mappers.Dataset).where(
-                Mappers.Dataset.dataset_id == dataset.dataset_id
-            )
-        )
-        assert mapped_dataset is not None
-        assert len(mapped_dataset.reactions) == expected_count
-
-
-def test_main_parquet_rejects_changed_without_overwrite(prepared_engine, tmp_path):
-    """A re-ingest with mutated content but no --overwrite raises."""
-    parquet_path, dataset = _write_parquet_dataset(tmp_path)
-    argv = ["--dsn", str(prepared_engine.url), "--pattern", parquet_path]
-    add_datasets.main(add_datasets.parse_args(argv))
-    dataset.reactions[0].outcomes[0].conversion.value = 999.0
-    parquet.save_dataset(dataset, parquet_path)
-    with pytest.raises(
-        RuntimeError
-    ):  # main() collects per-future failures and raises in aggregate.
-        add_datasets.main(add_datasets.parse_args(argv))
-
-
-def _derived_counts(engine) -> dict[str, int]:
-    with Session(engine) as session:
-        return {
-            table: session.execute(
-                text(f"SELECT count(*) FROM derived.{table}")  # noqa: S608  (constant)
-            ).scalar()
-            for table in (
-                "reaction_smiles",
-                "compound_smiles",
-                "product_compound_smiles",
-            )
-        }
-
-
-def test_stages_ingest_then_derived(prepared_engine):
-    """--stages=ingest writes ord.*/public.* with no derived rows; a later --stages=derived backfills them."""
-    base = ["--dsn", str(prepared_engine.url), "--pattern", _PBTXT_FIXTURE]
-    add_datasets.main(add_datasets.parse_args([*base, "--stages", "ingest"]))
     with Session(prepared_engine) as session:
         assert session.query(Mappers.Reaction).count() > 0
-    assert _derived_counts(prepared_engine) == {
-        "reaction_smiles": 0,
-        "compound_smiles": 0,
-        "product_compound_smiles": 0,
-    }
-    add_datasets.main(add_datasets.parse_args([*base, "--stages", "derived"]))
-    counts = _derived_counts(prepared_engine)
-    assert counts["reaction_smiles"] > 0
-    with Session(prepared_engine) as session:
-        assert (
-            session.execute(text("SELECT count(*) FROM rdkit.reactions")).scalar() > 0
-        )
 
 
 def test_unknown_stage_rejected():
     with pytest.raises(SystemExit):
         add_datasets.main(
             add_datasets.parse_args(["--pattern", _PBTXT_FIXTURE, "--stages", "bogus"])
+        )
+
+
+def test_classify_reactions_requires_extra(monkeypatch):
+    """--classify_reactions exits early with a clear message when the extra is missing."""
+    monkeypatch.setattr(add_datasets.database, "update_reaction_classes", None)
+    with pytest.raises(SystemExit, match="reaction-class"):
+        add_datasets.main(
+            add_datasets.parse_args(
+                ["--pattern", _PBTXT_FIXTURE, "--classify_reactions"]
+            )
         )
